+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/SpringDataNeo4jResourceAdapter.kt
@@ -10,6 +10,7 @@ import eu.tib.orkg.prototype.statements.domain.model.ResourceId
 import eu.tib.orkg.prototype.statements.spi.ResourceRepository
 import eu.tib.orkg.prototype.statements.spi.ResourceRepository.ResourceContributors
 import java.util.*
+import kotlin.streams.asSequence
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Component
@@ -33,7 +34,8 @@ class SpringDataNeo4jResourceAdapter(
         neo4jRepository.deleteAll()
     }
 
-    override fun findAll(): Iterable<Resource> = neo4jRepository.findAll().map(Neo4jResource::toResource)
+    // The function name is a bit odd, but findAll() conflicts with pre-defined methods. This is documented to work.
+    override fun findAll(): Sequence<Resource> = neo4jRepository.findAllBy().asSequence().map(Neo4jResource::toResource)
 
     override fun findAll(pageable: Pageable): Page<Resource> =
         neo4jRepository.findAll(pageable).map(Neo4jResource::toResource)