+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/internal/Neo4jPredicateRepository.kt
@@ -2,6 +2,7 @@ package eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal
 
 import eu.tib.orkg.prototype.statements.domain.model.PredicateId
 import java.util.Optional
+import java.util.stream.Stream
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.neo4j.repository.Neo4jRepository
@@ -9,6 +10,9 @@ import org.springframework.data.neo4j.repository.Neo4jRepository
 interface Neo4jPredicateRepository : Neo4jRepository<Neo4jPredicate, Long> {
     @Deprecated("Migrate to the pageable version.")
     override fun findAll(): Iterable<Neo4jPredicate>
+
+    // The function name is a bit odd, but findAll() conflicts with pre-defined methods. This is documented to work.
+    fun findAllBy(): Stream<Neo4jPredicate>
 
     override fun findAll(pageable: Pageable): Page<Neo4jPredicate>