+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/internal/Neo4jStatement.kt
@@ -2,8 +2,6 @@ package eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import eu.tib.orkg.prototype.contributions.domain.model.ContributorId
-import eu.tib.orkg.prototype.escapeLiterals
-import eu.tib.orkg.prototype.statements.application.rdf.RdfConstants
 import eu.tib.orkg.prototype.statements.domain.model.PredicateId
 import eu.tib.orkg.prototype.statements.domain.model.StatementId
 import eu.tib.orkg.prototype.statements.domain.model.neo4j.AuditableEntity
@@ -64,30 +62,5 @@ data class Neo4jStatement(
 
     override fun toString(): String {
         return "{id:$statementId}==(${subject!!.thingId} {${subject!!.label}})-[$predicateId]->(${`object`!!.thingId} {${`object`!!.label}})=="
-    }
-
-    /**
-     * Convert the triple to a statement in NTriple format.
-     */
-    fun toNTriple(): String {
-        val pPrefix = RdfConstants.PREDICATE_NS
-        val result = "${serializeThing(subject!!)} <$pPrefix$predicateId> ${serializeThing(`object`!!)} ."
-        if (result[0] == '"')
-            // Ignore literal
-            // TODO: log this somewhere
-            return ""
-        return result
-    }
-
-    private fun serializeThing(thing: Neo4jThing): String {
-        val rPrefix = RdfConstants.RESOURCE_NS
-        val pPrefix = RdfConstants.PREDICATE_NS
-        val cPrefix = RdfConstants.CLASS_NS
-        return when (thing) {
-            is Neo4jResource -> "<$rPrefix${thing.thingId}>"
-            is Neo4jPredicate -> "<$pPrefix${thing.thingId}>"
-            is Neo4jClass -> "<$cPrefix${thing.thingId}>"
-            else -> "\"${escapeLiterals(thing.label!!)}\"^^<http://www.w3.org/2001/XMLSchema#string>"
-        }
     }
 }