+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/Class.kt
@@ -3,6 +3,8 @@ package eu.tib.orkg.prototype.statements.domain.model
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonProperty
 import eu.tib.orkg.prototype.contributions.domain.model.ContributorId
+import eu.tib.orkg.prototype.escapeLiterals
+import eu.tib.orkg.prototype.statements.application.rdf.RdfConstants
 import java.net.URI
 import java.time.OffsetDateTime
 import java.util.*
@@ -28,3 +30,13 @@ data class Class(
 }
 
 internal fun Class?.toOptional() = Optional.ofNullable(this)
+
+fun Class.toNTriple(): String {
+    val cPrefix = RdfConstants.CLASS_NS
+    val sb = StringBuilder()
+    sb.append("<$cPrefix$id> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.w3.org/2002/07/owl#Class> .\n")
+    if (uri != null)
+        sb.append("<$cPrefix$id> <http://www.w3.org/2002/07/owl#equivalentClass> <$uri> .\n")
+    sb.append("<$cPrefix$id> <http://www.w3.org/2000/01/rdf-schema#label> \"${escapeLiterals(label)}\"^^<http://www.w3.org/2001/XMLSchema#string> .\n")
+    return sb.toString()
+}