+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/Resource.kt
@@ -3,7 +3,10 @@ package eu.tib.orkg.prototype.statements.domain.model
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonProperty
 import eu.tib.orkg.prototype.contributions.domain.model.ContributorId
+import eu.tib.orkg.prototype.escapeLiterals
 import eu.tib.orkg.prototype.statements.application.ExtractionMethod
+import eu.tib.orkg.prototype.statements.application.rdf.RdfConstants
+import java.lang.StringBuilder
 import org.eclipse.rdf4j.model.Model
 import java.time.OffsetDateTime
 
@@ -32,4 +35,14 @@ data class Resource(
 ) : Thing {
     @JsonIgnore
     var rdf: Model? = null
+}
+
+fun Resource.toNTriple(): String {
+    val cPrefix = RdfConstants.CLASS_NS
+    val rPrefix = RdfConstants.RESOURCE_NS
+    val sb = StringBuilder()
+    sb.append("<$rPrefix$id> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <${cPrefix}Resource> .\n")
+    classes.forEach { sb.append("<$rPrefix$id> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <$cPrefix${it.value}> .\n") }
+    sb.append("<$rPrefix$id> <http://www.w3.org/2000/01/rdf-schema#label> \"${escapeLiterals(label)}\"^^<http://www.w3.org/2001/XMLSchema#string> .\n")
+    return sb.toString()
 }