+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/rdf/RdfService.kt
@@ -1,5 +1,7 @@
 package eu.tib.orkg.prototype.statements.domain.model.rdf
 
+import java.io.OutputStream
+
 interface RdfService {
-    fun dumpToNTriple(): String
+    fun dumpToNTriple(out: OutputStream)
 }