+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/infrastructure/neo4j/rdf/Neo4jRdfService.kt
@@ -1,47 +1,33 @@
 package eu.tib.orkg.prototype.statements.infrastructure.neo4j.rdf
 
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jClass
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jClassRepository
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jResource
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jResourceRepository
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jStatement
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jStatementRepository
+import eu.tib.orkg.prototype.statements.domain.model.Class
+import eu.tib.orkg.prototype.statements.domain.model.GeneralStatement
 import eu.tib.orkg.prototype.statements.domain.model.Predicate
+import eu.tib.orkg.prototype.statements.domain.model.Resource
 import eu.tib.orkg.prototype.statements.domain.model.rdf.RdfService
 import eu.tib.orkg.prototype.statements.domain.model.toNTriple
+import eu.tib.orkg.prototype.statements.spi.ClassRepository
 import eu.tib.orkg.prototype.statements.spi.PredicateRepository
-import org.springframework.data.domain.Page
-import org.springframework.data.domain.PageRequest
+import eu.tib.orkg.prototype.statements.spi.ResourceRepository
+import eu.tib.orkg.prototype.statements.spi.StatementRepository
+import java.io.OutputStream
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 @Transactional
 class Neo4jRdfService(
-    private val statementRepository: Neo4jStatementRepository,
+    private val statementRepository: StatementRepository,
     private val predicateRepository: PredicateRepository,
-    private val resourceRepository: Neo4jResourceRepository,
-    private val classesRepository: Neo4jClassRepository
+    private val resourceRepository: ResourceRepository,
+    private val classesRepository: ClassRepository
 ) : RdfService {
-    override fun dumpToNTriple(): String {
-        // dump classes
-        var result = classesRepository.findAll().joinToString("\n", transform = Neo4jClass::toNTriple)
-        result += "\n"
-        // dump predicates
-        var predicates: Page<Predicate> = predicateRepository.findAll(PageRequest.of(0, 1000))
-        result += predicates.joinToString("\n", transform = Predicate::toNTriple)
-        result += "\n"
-        while (predicates.hasNext()) {
-            predicates = predicateRepository.findAll(predicates.nextPageable())
-            result += predicates.joinToString("\n", transform = Predicate::toNTriple)
-            result += "\n"
-        }
-        // dump resources
-        result += resourceRepository.findAll().joinToString("\n", transform = Neo4jResource::toNTriple)
-        result += "\n"
-        // dump statements
-        result += statementRepository.findAll().joinToString("\n", transform = Neo4jStatement::toNTriple)
-        result += "\n"
-        return result
+    override fun dumpToNTriple(out: OutputStream) {
+        val everything = sequenceOf<String>() + // Just for aligning the code below…
+            classesRepository.findAll().map(Class::toNTriple) +
+            predicateRepository.findAll().map(Predicate::toNTriple) +
+            resourceRepository.findAll().map(Resource::toNTriple) +
+            statementRepository.findAll().map(GeneralStatement::toNTriple)
+        everything.forEach { out.write(it.toByteArray()) }
     }
 }